span inner class
