+++ data/src/main/java/com/microsoft/azure/kusto/data/KustoOperationResult.java
@@ -11,7 +11,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.microsoft.azure.kusto.data.exceptions.JsonPropertyMissingException;
 import com.microsoft.azure.kusto.data.exceptions.KustoServiceQueryError;
-import com.microsoft.azure.kusto.data.instrumentation.KustoSpan;
 import com.microsoft.azure.kusto.data.instrumentation.KustoTracer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,7 +41,7 @@ public class KustoOperationResult implements Iterator<KustoResultSetTable> {
 
     public KustoOperationResult(String response, String version) throws KustoServiceQueryError {
 
-        KustoSpan kustoSpan = KustoTracer.startSpan("KustoOperationResult.createFromResponse", Context.NONE, ProcessKind.PROCESS, null);
+        KustoTracer.KustoSpan kustoSpan = KustoTracer.startSpan("KustoOperationResult.createFromResponse", Context.NONE, ProcessKind.PROCESS, null);
         try (kustoSpan){
             if (version.contains("v2")) {
                 createFromV2Response(response);
