+++ data/src/main/java/com/microsoft/azure/kusto/data/auth/TokenProviderBase.java
@@ -9,14 +9,12 @@ import java.net.URISyntaxException;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.microsoft.azure.kusto.data.instrumentation.KustoSpan;
 import com.microsoft.azure.kusto.data.instrumentation.KustoTracer;
 import org.apache.http.client.HttpClient;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import sun.font.EAttribute;
 
 public abstract class TokenProviderBase {
     protected final Logger logger = LoggerFactory.getLogger(getClass());
@@ -32,7 +30,7 @@ public abstract class TokenProviderBase {
         // trace GetToken
         Map<String, String> attributes = new HashMap<>();
         attributes.put("authentication_method", getAuthMethodForTracing());
-        KustoSpan kustoSpan = KustoTracer.startSpan("TokenProvider.acquireAccessToken", Context.NONE, ProcessKind.PROCESS, attributes);
+        KustoTracer.KustoSpan kustoSpan = KustoTracer.startSpan("TokenProvider.acquireAccessToken", Context.NONE, ProcessKind.PROCESS, attributes);
         try (kustoSpan){
             return acquireAccessTokenInner();
         } catch (DataServiceException | DataClientException e){
