+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/IngestClientBase.java
@@ -2,7 +2,6 @@ package com.microsoft.azure.kusto.ingest;
 
 import com.azure.core.util.Context;
 import com.azure.core.util.tracing.ProcessKind;
-import com.microsoft.azure.kusto.data.instrumentation.KustoSpan;
 import com.microsoft.azure.kusto.data.instrumentation.KustoTracer;
 import com.microsoft.azure.kusto.ingest.exceptions.IngestionClientException;
 import com.microsoft.azure.kusto.ingest.exceptions.IngestionServiceException;
@@ -45,7 +44,7 @@ public abstract class IngestClientBase implements IngestClient{
         // trace ingestFromFile
         Map<String, String> attributes = new HashMap<>();
         attributes.put("ingestFromFile", "complete");
-        KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.ingestFromFile", Context.NONE, ProcessKind.PROCESS, attributes);
+        KustoTracer.KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.ingestFromFile", Context.NONE, ProcessKind.PROCESS, attributes);
         try (kustoSpan) {
             return ingestFromFileImpl(fileSourceInfo, ingestionProperties);
         } catch (IngestionClientException | IngestionServiceException e) {
@@ -76,7 +75,7 @@ public abstract class IngestClientBase implements IngestClient{
         // trace ingestFromBlob
         Map<String, String> attributes = new HashMap<>();
         attributes.put("ingestFromBlob", "complete");
-        KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.ingestFromBlob", Context.NONE, ProcessKind.PROCESS, attributes);
+        KustoTracer.KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.ingestFromBlob", Context.NONE, ProcessKind.PROCESS, attributes);
         try (kustoSpan){
             return ingestFromBlobImpl(blobSourceInfo, ingestionProperties);
         } catch (IngestionClientException | IngestionServiceException e){
@@ -110,7 +109,7 @@ public abstract class IngestClientBase implements IngestClient{
         // trace ingestFromResultSet
         Map<String, String> attributes = new HashMap<>();
         attributes.put("ingestFromResultSet", "complete");
-        KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.ingestFromResultSet", Context.NONE, ProcessKind.PROCESS, attributes);
+        KustoTracer.KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.ingestFromResultSet", Context.NONE, ProcessKind.PROCESS, attributes);
         try (kustoSpan){
             return ingestFromResultSetImpl(resultSetSourceInfo, ingestionProperties);
         } catch (IngestionClientException | IngestionServiceException e){
@@ -141,7 +140,7 @@ public abstract class IngestClientBase implements IngestClient{
         // trace ingestFromStream
         Map<String, String> attributes = new HashMap<>();
         attributes.put("ingestFromStream", "complete");
-        KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.", Context.NONE, ProcessKind.PROCESS, attributes);
+        KustoTracer.KustoSpan kustoSpan = KustoTracer.startSpan("IngestClientBase.", Context.NONE, ProcessKind.PROCESS, attributes);
         try (kustoSpan){
             return ingestFromStreamImpl(streamSourceInfo, ingestionProperties);
         } catch (IngestionClientException | IngestionServiceException e){
